// client/src/__generated__/ChangePasswordMutation.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL mutation operation: ChangePasswordMutation
// ====================================================

export interface ChangePasswordMutation {
  changePassword: boolean;
}

export interface ChangePasswordMutationVariables {
  oldPassword: string;
  newPassword: string;
}